transpile: Split off new `macros` module
